requirements.txt
streamlit==1.50.0
Pillow==11.3.0
boto3==1.40.52
python-dotenv==1.1.1
PyYAML==6.0.3
numpy==2.2.6
faiss-cpu==1.12.0
PyPDF2==3.0.1
pdfplumber==0.11.7
reportlab==4.4.4
semantic-kernel==1.37.0
